Create VoiceGPT-ElevenLabs
This python code creates a voice enabled ChatGPT with OpenAI.
User will ask questions using microphone, it will be transcribed by Whisper API, sent to the GPT3.5 model and receive response. The response will then be converted to speech using Eleven Labs api as audio stream and be played back using ffplay.
The UI is generated by Gradio and the conversation transcript will be displayed on the screen.